+++ packages/plugin-table/web/src/icons/TrashIcon.tsx
@@ -8,6 +8,7 @@ const TrashIcon = props => (
     height={16}
     viewBox="0 0 19 19"
     {...props}
+    fill="currentColor"
   >
     <defs>
       <path