+++ packages/ui-components/web/src/Icons/ReplaceIcon.tsx
@@ -14,7 +14,7 @@ const ReplaceIcon = (props: any) => (
       clipRule="evenodd"
       // eslint-disable-next-line max-len
       d="M5.7 9H8V10H4V6H5V8.1C6.4 5.6 9.1 4 12 4C16.4 4 20 7.6 20 12H19C19 8.1 15.9 5 12 5C9.3 5 6.8 6.6 5.7 9ZM18.3 15H16V14H20V18H19V15.9C17.6 18.4 14.9 20 12 20C7.6 20 4 16.4 4 12H5C5 15.9 8.1 19 12 19C14.7 19 17.1 17.4 18.3 15Z"
-      fill="black"
+      fill="currentColor"
     />
   </svg>
 );